WIP - Adding login screen and default components for UIs
