+++ kp-commons-root/kp-commons-ui/src/main/java/de/kaiserpfalzedv/commons/webui/ui/KPOfficeUI.java
@@ -14,32 +14,31 @@
  * limitations under the License.
  */
 
-package de.kaiserpfalzedv.office.webui.ui;
+    package de.kaiserpfalzedv.commons.webui.ui;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.context.SessionScoped;
-import javax.inject.Inject;
+    import com.vaadin.annotations.Push;
+    import com.vaadin.annotations.Theme;
+    import com.vaadin.annotations.Viewport;
+    import com.vaadin.cdi.CDIUI;
+    import com.vaadin.cdi.access.AccessControl;
+    import com.vaadin.navigator.Navigator;
+    import com.vaadin.navigator.ViewProvider;
+    import com.vaadin.server.Responsive;
+    import com.vaadin.server.VaadinRequest;
+    import com.vaadin.ui.CssLayout;
+    import com.vaadin.ui.HorizontalLayout;
+    import com.vaadin.ui.UI;
+    import com.vaadin.ui.themes.ValoTheme;
+    import de.kaiserpfalzedv.vaadin.menu.Menu;
+    import org.vaadin.addon.cdiproperties.TextBundle;
+    import org.vaadin.addon.cdiproperties.annotation.CssLayoutProperties;
+    import org.vaadin.addon.cdiproperties.annotation.HorizontalLayoutProperties;
 
-import com.vaadin.annotations.Push;
-import com.vaadin.annotations.Theme;
-import com.vaadin.annotations.Viewport;
-import com.vaadin.annotations.Widgetset;
-import com.vaadin.cdi.CDIUI;
-import com.vaadin.cdi.access.AccessControl;
-import com.vaadin.navigator.Navigator;
-import com.vaadin.navigator.ViewProvider;
-import com.vaadin.server.Responsive;
-import com.vaadin.server.VaadinRequest;
-import com.vaadin.ui.CssLayout;
-import com.vaadin.ui.HorizontalLayout;
-import com.vaadin.ui.UI;
-import com.vaadin.ui.themes.ValoTheme;
-import de.kaiserpfalzedv.vaadin.menu.Menu;
-import org.vaadin.addon.cdiproperties.TextBundle;
-import org.vaadin.addon.cdiproperties.annotation.CssLayoutProperties;
-import org.vaadin.addon.cdiproperties.annotation.HorizontalLayoutProperties;
+    import javax.enterprise.context.ApplicationScoped;
+    import javax.enterprise.context.SessionScoped;
+    import javax.inject.Inject;
 
-import static org.apache.commons.lang3.StringUtils.isNotBlank;
+    import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 @Viewport("user-scalable=no,initial-scale=1.0")
 @Push
