+++ kp-commons-root/kp-commons-ui/src/main/java/de/kaiserpfalzedv/commons/webui/ui/login/LoginScreenView.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Kaiserpfalz EDV-Service, Roland T. Lichti
+ * Copyright 2017 Kaiserpfalz EDV-Service, Roland T. Lichti
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,17 @@
  * limitations under the License.
  */
 
-package de.kaiserpfalzedv.office.webui.ui.login;
+package de.kaiserpfalzedv.commons.webui.ui.login;
 
-import java.io.Serializable;
+import org.vaadin.addon.cdimvp.MVPView;
 
 /**
  * @author klenkes {@literal <rlichti@kaiserpfalz-edv.de>}
  * @version 1.0.0
- * @since 2016-07-10
+ * @since 2016-07-03
  */
-public class ForgottPasswordEvent implements Serializable {
+public interface LoginScreenView extends MVPView {
+    String getUserName();
+
+    String getPassword();
 }
