+++ kp-finance-root/kp-finance-accounting-ui/pom.xml
@@ -31,7 +31,7 @@
     <artifactId>kp-finance-accounting-ui</artifactId>
     <packaging>war</packaging>
 
-    <name>KP Finance :: Accounting UI</name>
+    <name>KP Finance :: UI :: Accounting</name>
 
     <properties>
         <vaadin.widgetset.mode>local</vaadin.widgetset.mode>
